Migrated the "algorithm libraries" to traccc_add_library.
At the same time moving all public headers under "traccc/".
