+++ examples/algorithms/cpu/include/traccc/clusterization/clusterization_algorithm.hpp
@@ -1,23 +1,23 @@
 /** TRACCC library, part of the ACTS project (R&D line)
  *
- * (c) 2021 CERN for the benefit of the ACTS project
+ * (c) 2021-2022 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
 
 #pragma once
 
-#include "edm/cell.hpp"
-#include "edm/cluster.hpp"
-#include "edm/internal_spacepoint.hpp"
-#include "edm/measurement.hpp"
-#include "edm/spacepoint.hpp"
-#include "geometry/pixel_segmentation.hpp"
+#include "traccc/edm/cell.hpp"
+#include "traccc/edm/cluster.hpp"
+#include "traccc/edm/internal_spacepoint.hpp"
+#include "traccc/edm/measurement.hpp"
+#include "traccc/edm/spacepoint.hpp"
+#include "traccc/geometry/pixel_segmentation.hpp"
 
 // clusterization
-#include "clusterization/component_connection.hpp"
-#include "clusterization/measurement_creation.hpp"
-#include "clusterization/spacepoint_formation.hpp"
+#include "traccc/clusterization/component_connection.hpp"
+#include "traccc/clusterization/measurement_creation.hpp"
+#include "traccc/clusterization/spacepoint_formation.hpp"
 
 #ifdef _OPENMP
 #include "omp.h"
