+++ examples/algorithms/cuda/CMakeLists.txt
@@ -1,12 +1,11 @@
-add_library(traccc_algorithms_cuda INTERFACE)
+# TRACCC library, part of the ACTS project (R&D line)
+#
+# (c) 2021-2022 CERN for the benefit of the ACTS project
+#
+# Mozilla Public License Version 2.0
 
-target_include_directories(traccc_algorithms_cuda
-  INTERFACE
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include>
-)
-
-target_link_libraries(traccc_algorithms_cuda
-  INTERFACE traccc::io traccc::core traccc::cuda)
-
-add_library(traccc::algorithms_cuda ALIAS traccc_algorithms_cuda)
+# Set up the "build" of the traccc::algorithms_cuda library.
+traccc_add_library( traccc_algorithms_cuda algorithms_cuda TYPE INTERFACE
+  "include/traccc/cuda/track_finding/seeding_algorithm.hpp" )
+target_link_libraries( traccc_algorithms_cuda
+  INTERFACE vecmem::core traccc::core traccc::cuda )
