+++ examples/algorithms/sycl/CMakeLists.txt
@@ -1,7 +1,11 @@
-add_library(traccc_algorithms_sycl INTERFACE
-${CMAKE_CURRENT_SOURCE_DIR}/include/sycl/track_finding/seeding_algorithm.hpp)
+# TRACCC library, part of the ACTS project (R&D line)
+#
+# (c) 2022 CERN for the benefit of the ACTS project
+#
+# Mozilla Public License Version 2.0
 
-target_link_libraries(traccc_algorithms_sycl
-  INTERFACE traccc::io traccc::core traccc::sycl)
-
-add_library(traccc::algorithms_sycl ALIAS traccc_algorithms_sycl) 
+# Set up the "build" of the traccc::algorithms_sycl library.
+traccc_add_library( traccc_algorithms_sycl algorithms_sycl TYPE INTERFACE
+  "include/traccc/sycl/track_finding/seeding_algorithm.hpp" )
+target_link_libraries( traccc_algorithms_sycl
+  INTERFACE vecmem::core traccc::core traccc::sycl )
